requirements.txt
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.2
Pillow==10.4.0
opencv-python==4.10.0.84
scikit-image==0.24.0
matplotlib==3.8.4
joblib==1.4.2
tqdm==4.66.4
